moves make_charge_fields to utils
